@@ -4,11 +4,11 @@
     "homepage": "https://github.com/karsonzhang/fastadmin-addons",
     "license": "Apache-2.0",
     "minimum-stability": "dev",
-    "version": "1.1.10",
+    "version": "1.1.11",
     "authors": [
         {
-            "name": "karsonzhang",
-            "email": "karsonzhang@163.com"
+            "name": "Karson",
+            "email": "karson@fastadmin.net"
         },
         {
             "name": "xiaobo.sun",
@@ -19,7 +19,7 @@
         "issues": "https://github.com/karsonzhang/fastadmin-addons/issues"
     },
     "require": {
-        "php": ">=5.6.0"
+        "php": ">=7.0.0"
     },
     "autoload": {
         "psr-4": {
