@@ -1269,7 +1269,106 @@ where cpev.attribute_id in(161,163,164) and cpev.store_id=0 and cpev.entity_id=$
         ini_set('memory_limit', '512M');
 
         $str = [
-            
+            400303475,
+            400302697,
+            400302989,
+            400302959,
+            400304649,
+            400302940,
+            500007412,
+            500007417,
+            400302964,
+            400302894,
+            400303050,
+            400302708,
+            400302839,
+            400302902,
+            400303398,
+            400289655,
+            400303016,
+            400304579,
+            400302702,
+            400302840,
+            400303734,
+            400302900,
+            400302991,
+            100137394,
+            400303588,
+            400303638,
+            400302781,
+            400303002,
+            400302551,
+            400302920,
+            400302829,
+            400303680,
+            400303041,
+            500007430,
+            400303602,
+            100137089,
+            400302932,
+            400303559,
+            400302845,
+            400303657,
+            400302811,
+            400304592,
+            100137097,
+            400303046,
+            400302795,
+            500007482,
+            100137383,
+            400303786,
+            500007428,
+            400303706,
+            400305086,
+            400304501,
+            400303027,
+            400302698,
+            400304610,
+            400304618,
+            400303777,
+            400303773,
+            400303480,
+            400303485,
+            400302836,
+            400302834,
+            400304550,
+            400304505,
+            400304593,
+            500007422,
+            400302965,
+            100137389,
+            400302565,
+            400302685,
+            500007465,
+            400302922,
+            400303020,
+            400304646,
+            100137172,
+            400304514,
+            500007395,
+            400304625,
+            400304445,
+            400303770,
+            400303781,
+            400304524,
+            400302710,
+            400303731,
+            400304464,
+            400303269,
+            100137390,
+            100137367,
+            400303687,
+            400303583,
+            400305084,
+            400303369,
+            400303537,
+            400303738,
+            400288529,
+            400303614,
+            100137402,
+            100137411,
+            100137520,
+            400303608
         ];
 
         $map['sfo.increment_id'] = ['in', $str];
