@@ -679,6 +679,14 @@ return [
         32,
         33,
         34,
+        96,
+        97,
+        98,
+        99,
+        100,
+        101,
+        102,
+        103,
         36
     ],
     //仓库部门角色组ID
@@ -724,40 +732,37 @@ return [
     'manage_time_out' => 86400,
     //措施里面超时配置
     'step_time_out' =>[
-        
         1 => 60*60*24,
         2 => 60*60*48,
         3 => 60*60*24,
-        4 => 60*60*72,
+        4 => 60*60*24,
         6 => 60*60*24,
-        7 => 60*60*24,
-        9 => 60*60*48,
-        10 => 60*60*48
+        11 => 60*60*72
     ],
     //仓库问题类型的超时配置
     'warehouse_time_out' =>[
-        // 1 => '核实处方',
-        // 2 => '不可加工',
-        // 3 => '镜框缺货',
-        // 4 => '镜片缺货',
-        // 5 => '镜片重做',
-        // 6 => '定制片超时',
-        // 7 => '核实地址',
-        // 8 => '系统漏单',
-        // 9 => '配错镜框',
-        // 10 => '物流退件',
-        // 11 => '客户退件',
-        1=> 60*60*72,
-        2=> 60*60*72,
-        3=> 60*60*72,
-        4=> 60*60*12,
-        5=> 60*60*12,
-        6=> 60*60*12,
-        7=> 60*60*72,
-        8=> 60*60*12,
-        9=> 60*60*12,
-        10=>60*60*72,
-        11=>60*60*72,
+        // 26 => '镜框缺货',
+        // 27 => '镜片缺货',
+        // 28 => '镜片重做',
+        // 29 => '定制片超时',
+        // 25 => '不可加工',
+        // 24 => '核实处方',
+        // 30 => '核实地址',
+        // 31 => '系统漏单',
+        // 32 => '配错镜框',
+        // 33 => '物流退件',
+        // 34 => '客户退件',
+        26=> 60*60*72,
+        27=> 60*60*12,
+        28=> 60*60*12,
+        29=> 60*60*12,
+        25=> 60*60*72,
+        24=> 60*60*72,
+        30=> 60*60*72,
+        31=> 60*60*12,
+        32=> 60*60*12,
+        33=>60*60*72,
+        34=>60*60*72,
     ],
     //客服数据统计是筛选的分组 A/B组
     'customer_type' =>[
@@ -778,5 +783,14 @@ return [
         2=>'电话组'
     ],
     //区分正式/试用期员工时间
-    'customer_category_time'=>60*60*24*60
+    'customer_category_time'=>60*60*24*60,
+    'cost_arr'=>[
+        ['type'=>'运费','money_us'=>10849.08,'money_cn'=>73773.74,'percent'=>'13.23'],
+        ['type'=>'镜架成本','money_us'=>7076.70,'money_cn'=>48121.54,'percent'=>'8.63'],
+        ['type'=>'镜片成本','money_us'=>0.00,'money_cn'=>0.00,'percent'=>'0.00'],
+        ['type'=>'Google Adwords','money_us'=>2664.64,'money_cn'=>18119.53,'percent'=>'3.25'],
+        ['type'=>'Facebook','money_us'=>29483.50,'money_cn'=>200487.80,'percent'=>'35.95'],
+        ['type'=>'毛利润','money_us'=>31938.88,'money_cn'=>217184.36,'percent'=>'38.94'],
+        ['type'=>'销售额','money_us'=>82012.79,'money_cn'=>557686.97,'percent'=>''],
+    ],
 ];
